Extract components from vehicle popup
