+++ frontend/src/app/shared/components/vehicle-load-unload-controls/vehicle-load-unload-controls.component.html
@@ -1,0 +1,21 @@
+<div class="btn-group">
+    <button
+        (click)="unloadVehicle()"
+        [disabled]="(vehicleLoadState$ | async) === 'completelyUnloaded'"
+        class="btn btn-outline-primary"
+        data-cy="vehiclePopupUnloadButton"
+    >
+        <i class="bi-box-arrow-left"></i>
+        Alle aussteigen
+    </button>
+    <button
+        (click)="loadVehicle()"
+        [disabled]="(vehicleLoadState$ | async) === 'completelyLoaded'"
+        class="btn btn-outline-primary"
+    >
+        <i class="bi-box-arrow-in-right"></i>
+        Alle einsteigen
+    </button>
+</div>
+
+<p>Personal, das sich gerade im Transfer befindet, kann nicht einsteigen.</p>
